refactor: Move allowlist logic to `migrate` mod
